requirements.txt
Django==5.1.5
tensorflow==2.20.0
Pillow==11.2.1
numpy==2.1.3
scikit-learn==1.6.1
djangorestframework==3.15.2
python-dotenv==1.0.1